:recycle: jwt parser generator 분리
