+++ src/main/kotlin/lifestyle/dt/global/filter/JwtTokenFilter.kt
@@ -1,6 +1,7 @@
 package lifestyle.dt.global.filter
 
-import lifestyle.dt.global.security.jwt.JwtTokenProvider
+import lifestyle.dt.global.security.jwt.JwtGenerator
+import lifestyle.dt.global.security.jwt.JwtParser
 import org.springframework.security.core.context.SecurityContextHolder
 import org.springframework.web.filter.OncePerRequestFilter
 import javax.servlet.FilterChain
@@ -8,7 +9,7 @@ import javax.servlet.http.HttpServletRequest
 import javax.servlet.http.HttpServletResponse
 
 class JwtTokenFilter(
-    private val jwtTokenProvider: JwtTokenProvider
+    private val jwtParser: JwtParser
 ): OncePerRequestFilter() {
 
     override fun doFilterInternal(
@@ -16,9 +17,9 @@ class JwtTokenFilter(
         response: HttpServletResponse,
         filterChain: FilterChain,
     ) {
-        val token: String? = jwtTokenProvider.resolveToken(request)
+        val token: String? = jwtParser.resolveToken(request)
         if (!token.isNullOrBlank()) {
-            val authentication = jwtTokenProvider.authentication(token)
+            val authentication = jwtParser.authentication(token)
             SecurityContextHolder.getContext().authentication = authentication
         }
         filterChain.doFilter(request, response)
