+++ src/main/kotlin/lifestyle/dt/global/security/jwt/JwtGenerator.kt
@@ -1,0 +1,48 @@
+package lifestyle.dt.global.security.jwt
+
+import io.jsonwebtoken.Jwts
+import io.jsonwebtoken.SignatureAlgorithm
+import lifestyle.dt.global.security.auth.AuthDetailsService
+import lifestyle.dt.global.security.jwt.properties.JwtProperties
+import org.springframework.stereotype.Component
+import java.security.Key
+import java.time.ZonedDateTime
+import java.util.*
+import javax.servlet.http.HttpServletRequest
+
+@Component
+class JwtGenerator(
+    private val jwtProperties: JwtProperties
+) {
+    companion object {
+        const val ACCESS_TYPE = "access"
+        const val REFRESH_TYPE = "refresh"
+        const val ACCESS_EXP = 60L * 15 // 15 min
+        const val REFRESH_EXP = 60L * 60 * 24 * 7 // 1 weeks
+        const val TOKEN_PREFIX = "Bearer "
+    }
+
+    val accessExpiredTime: ZonedDateTime
+        get() = ZonedDateTime.now().plusSeconds(ACCESS_EXP)
+
+    fun generateAccessToken(email: String): String =
+        generateToken(email, ACCESS_TYPE, jwtProperties.accessSecret, ACCESS_EXP)
+
+    fun generateRefreshToken(email: String): String =
+        generateToken(email, REFRESH_TYPE, jwtProperties.refreshSecret, REFRESH_EXP)
+
+
+    fun parseToken(token: String): String? =
+        if (token.startsWith(TOKEN_PREFIX)) token.replace(TOKEN_PREFIX, "") else null
+
+    fun generateToken(sub: String, type: String, secret: Key, exp: Long): String {
+        return Jwts.builder()
+            .signWith(secret, SignatureAlgorithm.HS256)
+            .setSubject(sub)
+            .claim("type", type)
+            .setIssuedAt(Date())
+            .setExpiration(Date(System.currentTimeMillis() + exp * 1000))
+            .compact()
+    }
+
+}
